ml-service/requirements.txt
# Core Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3

# Security and Authentication
bcrypt==4.0.1
Flask-Limiter==3.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0

# Machine Learning and Data Science
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.1
xgboost==1.7.6
lightgbm==4.0.0
tensorflow==2.13.0
keras==2.13.1

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Optimization and Algorithms
ortools==9.6.2534
scipy==1.11.1
pulp==2.7.0

# Database and Data Processing
psycopg2-binary==2.9.7
sqlalchemy==2.0.20

# API and HTTP
requests==2.31.0
gunicorn==21.2.0

# Logging and Monitoring
python-json-logger==2.0.7

# Additional Utilities
python-dotenv==1.0.0
click==8.1.6
